Fix/token precision

+++ FRW/Modules/Staking/StakeAmount/StakeAmountViewModel.swift
@@ -45,7 +45,7 @@ class StakeAmountViewModel: ObservableObject {
         self.isUnstake = isUnstake
         let token = WalletManager.shared.flowToken
         self.balance = isUnstake ? (provider.currentNode?.stakingCount ?? 0) : WalletManager.shared
-            .getBalance(byId: token?.contractId ?? "")
+            .getBalance(byId: token?.contractId ?? "").doubleValue
     }
 
     // MARK: Internal
